+++ functions-images.php
@@ -1,5 +1,7 @@
 <?php
 
+use Timmy\Timmy;
+
 if ( ! function_exists( 'get_timber_image' ) ) :
 /**
  * Outputs the src attr together with optional alt and title attributes
@@ -10,15 +12,14 @@ if ( ! function_exists( 'get_timber_image' ) ) :
  * @return string                           Src, alt and title attributes
  */
 function get_timber_image( $timber_image, $size ) {
-	// When we just have the post id, we convert it to a TimberImage
-	if ( is_numeric( $timber_image ) ) {
-		$timber_image = new Timber\Image( $timber_image );
-	}
-
 	$src  = get_timber_image_src( $timber_image, $size );
 	$attr = get_timber_image_attr( $timber_image );
 
-	return ' src="' . $src . '" ' . $attr;
+	if ( $src ) {
+		return ' src="' . $src . '" ' . $attr;
+	}
+
+	return '';
 }
 endif;
 
@@ -31,9 +32,10 @@ if ( ! function_exists( 'get_timber_image_src' ) ) :
  * @return string                           Image src
  */
 function get_timber_image_src( $timber_image, $size ) {
-	// When we just have the post id, we convert it to a TimberImage
-	if ( is_numeric( $timber_image ) ) {
-		$timber_image = new Timber\Image( $timber_image );
+	$timber_image = Timmy::get_timber_image( $timber_image );
+
+	if ( ! $timber_image ) {
+		return false;
 	}
 
 	$img_sizes = get_image_sizes();
@@ -81,6 +83,12 @@ if ( ! function_exists( 'get_timber_image_attr' ) ) :
  * @return string                       HTML string for alt and title attributes
  */
 function get_timber_image_attr( $timber_image ) {
+	$timber_image = Timmy::get_timber_image( $timber_image );
+
+	if ( ! $timber_image ) {
+		return '';
+	}
+
 	$alt   = $timber_image->_wp_attachment_image_alt;
 	$title = $timber_image->post_content;
 	return get_image_attr_html( $alt, $title );
@@ -96,15 +104,14 @@ if ( ! function_exists( 'get_timber_image_responsive' ) ) :
  * @return string                           Image srcset, sizes, alt and title attributes
  */
 function get_timber_image_responsive( $timber_image, $size ) {
-	// When we just have the post id, we convert it to a TimberImage
-	if ( is_numeric( $timber_image ) ) {
-		$timber_image = new Timber\Image( $timber_image );
-	}
-
 	$src = get_timber_image_responsive_src( $timber_image, $size );
 	$attr = get_timber_image_attr( $timber_image );
 
-	return $src . ' ' . $attr;
+	if ( $src ) {
+		return $src . ' ' . $attr;
+	}
+
+	return '';
 }
 endif;
 
@@ -117,9 +124,10 @@ if ( ! function_exists( 'get_timber_image_responsive_src' ) ) :
  * @return string                           Image srcset and sizes attributes
  */
 function get_timber_image_responsive_src( $timber_image, $size ) {
-	// When we just have the post id, we convert it to a TimberImage
-	if ( is_numeric( $timber_image ) ) {
-		$timber_image = new Timber\Image( $timber_image );
+	$timber_image = Timmy::get_timber_image( $timber_image );
+
+	if ( ! $timber_image ) {
+		return false;
 	}
 
 	$img_sizes = get_image_sizes();
@@ -220,12 +228,16 @@ if ( ! function_exists( 'get_timber_image_responsive_acf' ) ) :
  */
 function get_timber_image_responsive_acf( $name, $size ) {
 	$image = get_field( $name );
-	$timber_image = new Timber\Image( $image['id'] );
+	$timber_image = Timmy::get_timber_image( $image );
 
 	$src  = get_timber_image_responsive_src( $timber_image, $size );
 	$attr = get_acf_image_attr( $image );
 
-	return $src . ' ' . $attr;
+	if ( $src ) {
+		return $src . ' ' . $attr;
+	}
+
+	return '';
 }
 endif;