+++ init.php
@@ -1,0 +1,7 @@
+<?php
+/**
+ * This file is called when Timmy is autoloaded through composer.
+ * If Timmy installed as a plugin, then it will be loaded through timmy.php.
+ */
+use Timmy\Timmy;
+new Timmy();